[CHORE/#49] 패키지 수정
