+++ feature/src/main/java/com/terning/feature/home/changefilter/component/FilteringMainTitleText.kt
@@ -1,20 +1,21 @@
-package com.terning.feature.home.filtering.component
+package com.terning.feature.home.changefilter.component
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.terning.core.designsystem.theme.Grey300
+import androidx.compose.ui.tooling.preview.Preview
+import com.terning.core.designsystem.theme.Black
 import com.terning.core.designsystem.theme.TerningTheme
 
 @Composable
-fun FilteringSubTitleText(
+fun FilteringMainTitleText(
     text: String,
     modifier: Modifier = Modifier,
 ) {
     Text(
         text = text,
-        style = TerningTheme.typography.body5,
-        color = Grey300,
+        style = TerningTheme.typography.title3,
+        color = Black,
         modifier = modifier,
     )
 }
