+++ feature/src/main/java/com/terning/feature/home/changefilter/component/FilteringSubTitleText.kt
@@ -1,20 +1,21 @@
-package com.terning.feature.home.filtering.component
+package com.terning.feature.home.changefilter.component
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.terning.core.designsystem.theme.Black
+import androidx.compose.ui.tooling.preview.Preview
+import com.terning.core.designsystem.theme.Grey300
 import com.terning.core.designsystem.theme.TerningTheme
 
 @Composable
-fun FilteringMainTitleText(
+fun FilteringSubTitleText(
     text: String,
     modifier: Modifier = Modifier,
 ) {
     Text(
         text = text,
-        style = TerningTheme.typography.title3,
-        color = Black,
+        style = TerningTheme.typography.body5,
+        color = Grey300,
         modifier = modifier,
     )
 }
